requirements.txt
altgraph==0.17.5
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
beautifulsoup4==4.14.3
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
coverage==7.13.1
curl_cffi==0.13.0
dpkt==1.9.8
fastapi==0.128.0
feedparser==6.0.12
frozendict==2.4.7
greenlet==3.3.2
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.25.2
idna==3.11
iniconfig==2.3.0
loguru==0.7.2
lxml==6.0.2
multitasking==0.0.12
numpy==2.4.3
packaging==26.0
pandas==3.0.1
peewee==4.0.1
pefile==2024.8.26
platformdirs==4.9.4
pluggy==1.6.0
praw==7.8.1
prawcore==2.4.0
protobuf==7.34.0
psutil==5.9.6
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pyinstaller==6.19.0
pyinstaller-hooks-contrib==2026.1
PyQt6==6.10.2
PyQt6-Qt6==6.10.1
PyQt6_sip==13.11.0
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-json-logger==4.0.0
python-multipart==0.0.22
pytz==2026.1.post1
pywin32-ctypes==0.2.3
PyYAML==6.0.3
requests==2.31.0
schedule==1.2.2
scipy==1.17.1
setuptools==82.0.0
sgmllib3k==1.0.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.8.3
SQLAlchemy==2.0.48
starlette==0.50.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
update-checker==0.18.0
urllib3==2.6.3
uvicorn==0.40.0
watchfiles==1.1.1
websocket-client==1.9.0
websockets==16.0
win32_setctime==1.2.0
yfinance==1.2.0
